% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Dataset2}
\alias{Dataset2}
\title{Dataset without lurking variable}
\format{
A data frame with 500 rows and 3 variables:
\describe{
  \item{X}{A variable which is created out from Z}
  \item{Y}{The second varible which is created out from X}
  \item{Z}{The lurking variable, with is generated from the standard normal.}
}
}
\usage{
data(Dataset2)
}
\description{
A dataset containing three variables where X was made from Z and Y was made from X.
This ensure that Z is not a lurking variable.
}
\keyword{datasets}
