% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{earthquakes}
\alias{earthquakes}
\title{earthquakes}
\format{
A data frame with 107 rows and 2 variables:
\describe{
  \item{Year}{The year of the measurement}
  \item{n}{The number of major earthquakes measueret}
}
}
\usage{
data(earthquakes)
}
\description{
A dataset over the annual count of major earthquakes for the year 1900 to 2006.
Major earthquakes refer to earthquakes with a magnitude of 7 or above.
}
\keyword{datasets}
