% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Forward_algorithm_for_likelihood_calculation_of_ the_two-state
%   Bernoulli-HMM.R
\name{forwardalgo}
\alias{forwardalgo}
\title{Title}
\usage{
forwardalgo(delta, trans, emision, X, f)
}
\arguments{
\item{delta}{}

\item{trans}{}

\item{emision}{}

\item{X}{}

\item{f}{}
}
\value{

}
\description{
Title
}
\examples{
tpm = matrix(c(0.5,0.25,0.5,0.75),2,2)
del = c(1/3 , 2/3)
XX = c(1,1,1) 
emis = matrix(c(0.5,0,0.5,1),2,2) 
fbinom = function(x){x+1}
forwardalgo(delta = del,trans = tpm, emision = emis,X = XX,f=fbinom)
}
