% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMclass.R
\name{print.HMM}
\alias{print.HMM}
\title{print.HMM}
\usage{
print(HMM())
}
\arguments{
\item{HMM}{A HMM class object.}
}
\value{
A  HMM class object
}
\description{
This will print the descibtion of the model defined in the HMM object.
}
