% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viterbi.R
\name{viterbi}
\alias{viterbi}
\title{Viterbi algorithm}
\usage{
Viterbialgo(HMM,X)
}
\arguments{
\item{X}{Data}

\item{HMM}{A HMM object}
}
\value{
list() Returns a list of the most likely sequence of states.
}
\description{
Finds through dynamic programming the most likely sequence of states
}
\details{
D0es Global decoding through dynamic programming. It is not feasible to maximize over all possible states
because this would run in O(m^t). Thus Viterbi is used as an alternative to this approach.
}
