% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{earthquakes}
\alias{earthquakes}
\title{Dataset with lurking varible}
\format{
A data frame with 500 rows and 3 variables:
\describe{
  \item{X}{The first variable wich is created out from Z}
  \item{Y}{The second variable there is created out from Z}
}
}
\usage{
data(earthquakes)
}
\description{
A dataset containing two variables with are both generated from a model
dependted on the third variable Z.
}
\keyword{datasets}
