% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{em}
\alias{em}
\title{Expectation–Maximization (EM) Algorithm}
\usage{
em(HM,X)
}
\arguments{
\item{HM}{A HMM object}

\item{X}{Data}

\item{tol}{tolerance of percentage change between likelihoods from different iterations}

\item{maxiter}{the maximum number of iteration the algorithm will run}
}
\value{
HMM Returns an updated HMM object
}
\description{
The EM algorithm finds the maximum-likelihood estimates of a HMM.
}
\details{
Is an iterative method to find a local maximum likelihood or a maximum poosteriori estimates of parametersin a 
statistical model, where the model depends on hidden variables. The EM alternates between performing an expectation step, 
which creates a function for the expectation of the log-likelihood evaluated using the current estimate for the parameters, 
and a maximization step, which computes the parameters, which maximize the expected log-likelihood found in the E step. This 
is then iterated. The numbers of iteration the user want to do, are dependent on the tol and maxiter parameters. Maxiter is the maximum
number of iterations the user will allow the algorithm to run and tol is minimal percentage change in likelihood the algorithm need to see
before running another iterations. By using these two parameters, the user can control how long the algorithm should run. 

Note that if the user are using a distribution other then normal or poisson in the model, then the initial values need to be picked carefully. As 
numerical maximization is used to fine the optimal parameter, which can have problem with ill chosen starting parameters.
}
\examples{
 X <- earthquakes$n
delta = c(0.5,0.5)
trans=matrix(c(0.9,0.1,0.1,0.9),2,2)
HM = HMM(initial_dist = delta,transmission = trans,  
        emis_names = c("dpois","dpois"),parameterslist = list(list(lambda=10),list(lambda=30)) )
em(HM=HM,X=X)

}
