% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\title{Forward Algorithm}
\usage{
forward(HM,X,log_sum=F)
}
\arguments{
\item{HM}{A  HMM class object}

\item{X}{Data}

\item{log_sum}{True or False, If True then the log sum trick is used to deal with underflow.}
}
\value{
alpha_matrix A matrix of forward probabilities
}
\description{
Creates a vector of forward probabilities which is used in most of the functions in this package.
}
\details{
Does recursive computation of the likelihood which plays a key role in the likelihood evaluation
and thus parameter estimation. It is also used in forecasting, decoding and model checking. The recursive nature of
the forward algorithm is much more computationally inexpensive than brute-force summation over all possible state sequences.
}
