% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMclass.R
\name{HMM}
\alias{HMM}
\title{HMM}
\usage{
HMM(stationary_dist,transmission , emission_function, parameters)
}
\arguments{
\item{stationary_dist}{numerical vector of the stationary distribution}

\item{transmision}{numerical matrix containing the transmission probabilities}

\item{emission_function}{vector of emission functions}

\item{parameters}{list of list, where each list containing the parameters need in the emision functions}

\item{state_names}{vector of names for each state}
}
\value{
A  HMM class object
}
\description{
Will make a object containing the parameters of a hidden markov model.
}
