% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_prob.R
\name{state_prob}
\alias{state_prob}
\title{state_prob}
\usage{
state_prob(state,state_time,HMM,X)
}
\arguments{
\item{state}{}

\item{state_time}{}

\item{HM}{A HMM object}

\item{X}{Data}
}
\value{

}
\description{

}
