% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viterbi.R
\name{viterbi}
\alias{viterbi}
\title{Viterbi algorithm}
\usage{
viterbi(HM,X)
}
\arguments{
\item{HM}{A HMM object}

\item{X}{Data}
}
\value{
list() Returns a list of the most likely sequence of states.
}
\description{
Finds through dynamic programming the most likely sequence of states
}
\details{
Does Global decoding through dynamic programming. It is not feasible to maximize over all possible states
because this would run in O(m^t). Thus Viterbi is used as an alternative to this approach.
}
\examples{
#X = read.table("http://www.hmms-for-time-series.de/second/data/earthquakes.txt")[,2]
delta = c(0.5,0.5)
lambdaL=c(10,30)
trans=matrix(c(0.9,0.1,0.1,0.9),2,2)
hm = HMM(initial_dist = delta,transmission = trans,emis_names = c("dpois","dpois"),parameterslist = list(list(lambda =10),list(lambda =30)))
viterbi(hm,X)
}
